backend/policy_bot/requirements.txt
fastapi
uvicorn
python-dotenv
pdfplumber
langchain
faiss-cpu
google-generativeai==0.8.5
pymongo